refactor(tls): extract mod tests to sibling file
